--- DeviceTracking/app/src/main/java/com/alpsproject/devicetracking/ui/reports/PageViewModel.kt
@@ -1,19 +1,0 @@
-package com.alpsproject.devicetracking.ui.reports
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.Transformations
-import androidx.lifecycle.ViewModel
-
-class PageViewModel : ViewModel() {
-
-    private val _index = MutableLiveData<Int>()
-
-    val text: LiveData<String> = Transformations.map(_index) {
-        "Hello world from section: $it"
-    }
-
-    fun setIndex(index: Int) {
-        _index.value = index
-    }
-}
